build: Change build options for dev, production mode
